Clean up tables branch
