+++ oscen-lib/src/envelopes.rs
@@ -1,176 +1,157 @@
-use super::signal::*;
-use super::utils::ExpInterp;
-use crate::{as_any_mut, gate, std_signal};
-use std::{
-    any::Any,
-    ops::{Index, IndexMut},
-};
+use crate::rack::*;
+use crate::utils::{interp, interp_inv};
+use crate::{build, props, tag};
+use std::sync::Arc;
 
-#[derive(Copy, Clone)]
+#[derive(Copy, Clone, Debug)]
 pub struct Adsr {
     tag: Tag,
-    attack: In,
-    decay: In,
-    sustain: In,
-    release: In,
-    clock: Real,
-    sustain_time: Real,
-    triggered: bool,
-    level: Real,
-    a_param: Real,
-    d_param: Real,
-    r_param: Real,
-    a_interp: ExpInterp,
-    d_interp: ExpInterp,
-    r_interp: ExpInterp,
-    out: Real,
+    ax: f32,
+    dx: f32,
+    rx: f32,
 }
 
 impl Adsr {
-    pub fn new(id_gen: &mut IdGen, a_param: Real, d_param: Real, r_param: Real) -> Self {
-        let a_interp = ExpInterp::new(0.0, 0.5, 1.0);
-        let d_interp = ExpInterp::new(0.0, 0.5, 1.0);
-        let r_interp = ExpInterp::new(0.0, 0.5, 1.0);
+    pub fn new<T: Into<Tag>>(tag: T, ax: f32, dx: f32, rx: f32) -> Self {
         Self {
-            tag: id_gen.id(),
-            attack: (0.01).into(),
-            decay: 0.into(),
-            sustain: 1.into(),
-            release: (0.1).into(),
-            clock: 0.0,
-            sustain_time: 0.0,
-            triggered: false,
-            level: 0.0,
-            a_param,
-            d_param,
-            r_param,
-            a_interp,
-            d_interp,
-            r_interp,
-            out: 0.0,
+            tag: tag.into(),
+            ax,
+            dx,
+            rx,
         }
     }
+    props!(attack, set_attack, 0);
+    props!(decay, set_decay, 1);
+    props!(sustain, set_sustain, 2);
+    props!(release, set_release, 3);
 
-    pub fn linear(id_gen: &mut IdGen) -> Self {
-        Self::new(id_gen, 0.5, 0.5, 0.5)
-    }
-
-    pub fn exp_20(id_gen: &mut IdGen) -> Self {
-        Self::new(id_gen, 0.2, 0.2, 0.2)
-    }
-
-    pub fn attack<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.attack = arg.into();
-        self
-    }
-
-    pub fn decay<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.decay = arg.into();
-        self
-    }
-
-    pub fn sustain<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.sustain = arg.into();
-        self
-    }
-
-    pub fn release<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.release = arg.into();
-        self
-    }
-
-    pub fn calc_level(&mut self, rack: &Rack) -> Real {
-        fn max01(a: f64) -> f64 {
-            if a > 0.01 {
-                a
-            } else {
-                0.01
-            }
-        }
-
-        let a = max01(In::val(rack, self.attack));
-        let d = max01(In::val(rack, self.decay));
-        let s = In::val(rack, self.sustain);
-        let r = max01(In::val(rack, self.release));
-
-        if self.triggered {
-            match self.clock {
-                // Attack
-                t if t < a => self.a_interp.interp(t / a),
-                // Decay
-                t if t < a + d => self.d_interp.interp((t - a) / d),
-                // Sustain
-                t => {
-                    self.sustain_time = t - a - d;
-                    s
-                }
-            }
-        } else {
-            match self.clock {
-                // Attack
-                t if t < a => self.a_interp.interp(t / a),
-                // Decay
-                t if t < a + d => self.d_interp.interp((t - a) / d),
-                // Release
-                t if t < a + d + r + self.sustain_time => {
-                    self.r_interp.interp((t - a - d - self.sustain_time) / r)
-                }
-                // Off
-                _ => 0.,
-            }
+    pub fn triggered(&self, controls: &Controls) -> bool {
+        let ctrl = controls[(self.tag, 4)];
+        match ctrl {
+            Control::B(b) => b,
+            _ => panic!("triggered must be a bool, not {:?}", ctrl),
         }
     }
-
-    pub fn on(&mut self) {
-        self.triggered = true;
-        self.sustain_time = 0.0;
-        self.clock = self.a_interp.interp_inv(self.level);
+    pub fn set_triggered(&self, controls: &mut Controls, value: bool) {
+        controls[(self.tag, 4)] = value.into();
     }
 
-    pub fn off(&mut self) {
-        self.triggered = false;
+    pub fn on(&self, controls: &mut Controls, state: &mut State) {
+        self.set_triggered(controls, true);
+        state[(self.tag, 1)] = 0.0;
+        let x = state[(self.tag, 2)];
+        state[(self.tag, 0)] = interp_inv(0.0, 1.0 - self.ax, 1.0, x);
+    }
+    pub fn off(&self, controls: &mut Controls) {
+        self.set_triggered(controls, false);
     }
 }
-
-impl Builder for Adsr {}
-
-gate!(Adsr);
 
 impl Signal for Adsr {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, sample_rate: Real) -> Real {
-        self.a_interp.update(0.0, 1.0 - self.a_param, 1.0);
-        let s = In::val(rack, self.sustain);
-        self.d_interp.update(1.0, s + self.d_param * (1.0 - s), s);
-        self.r_interp.update(s, self.r_param * s, 0.0);
-        self.level = self.calc_level(rack);
-        self.clock += 1. / sample_rate;
-        self.level
+    tag!();
+    fn signal(
+        &self,
+        controls: &Controls,
+        state: &mut State,
+        outputs: &mut Outputs,
+        sample_rate: f32,
+    ) {
+        let a = self.attack(controls, outputs).max(0.005);
+        let d = self.decay(controls, outputs).max(0.005);
+        let s = self.sustain(controls, outputs);
+        let r = self.release(controls, outputs).max(0.005);
+        let triggered = self.triggered(controls);
+        state[(self.tag, 2)] = match (triggered, state[(self.tag, 0)]) {
+            (_, t) if t < a => interp(0.0, 1.0 - self.ax, 1.0, t / a),
+            (_, t) if t < a + d => interp(1.0, s + self.dx * (1.0 - s), s, (t - a) / d),
+            (true, t) => {
+                state[(self.tag, 1)] = t - a - d;
+                s
+            }
+            (false, t) if t < a + d + r + state[(self.tag, 1)] => {
+                interp(s, self.rx * s, 0.0, t - a - d - state[(self.tag, 1)] / r)
+            }
+            (false, _) => 0.0,
+        };
+        outputs[(self.tag, 0)] = state[(self.tag, 2)];
+        state[(self.tag, 0)] += 1.0 / sample_rate;
     }
 }
 
-impl Index<&str> for Adsr {
-    type Output = In;
-
-    fn index(&self, index: &str) -> &Self::Output {
-        match index {
-            "attack" => &self.attack,
-            "decay" => &self.decay,
-            "sustain" => &self.sustain,
-            "release" => &self.release,
-            _ => panic!("Adsr does not have a field named: {}", index),
-        }
-    }
+#[derive(Copy, Clone, Debug)]
+pub struct AdsrBuilder {
+    ax: f32,
+    dx: f32,
+    rx: f32,
+    attack: Control,
+    decay: Control,
+    sustain: Control,
+    release: Control,
+    triggered: Control,
 }
 
-impl IndexMut<&str> for Adsr {
-    fn index_mut(&mut self, index: &str) -> &mut Self::Output {
-        match index {
-            "attack" => &mut self.attack,
-            "decay" => &mut self.decay,
-            "sustain" => &mut self.sustain,
-            "release" => &mut self.release,
-            _ => panic!("Adsr does not have a field named: {}", index),
+impl AdsrBuilder {
+    pub fn new() -> Self {
+        let attack = 0.01.into();
+        let decay = 0.into();
+        let sustain = 1.into();
+        let release = 0.1.into();
+        let triggered = false.into();
+        Self {
+            ax: 0.5,
+            dx: 0.5,
+            rx: 0.5,
+            attack,
+            decay,
+            sustain,
+            release,
+            triggered,
         }
+    }
+    pub fn linear() -> Self {
+        let mut ab = Self::new();
+        ab.ax = 0.5;
+        ab.dx = 0.5;
+        ab.rx = 0.5;
+        ab
+    }
+    pub fn exp_20() -> Self {
+        let mut ab = Self::new();
+        ab.ax = 0.2;
+        ab.dx = 0.2;
+        ab.rx = 0.2;
+        ab
+    }
+    build!(attack);
+    build!(decay);
+    build!(sustain);
+    build!(release);
+
+    pub fn ax(&mut self, value: f32) -> &mut Self {
+        self.ax = value;
+        self
+    }
+    pub fn dx(&mut self, value: f32) -> &mut Self {
+        self.dx = value;
+        self
+    }
+    pub fn rx(&mut self, value: f32) -> &mut Self {
+        self.rx = value;
+        self
+    }
+    pub fn triggered(&mut self, t: bool) -> &mut Self {
+        self.triggered = t.into();
+        self
+    }
+    pub fn rack(&self, rack: &mut Rack, controls: &mut Controls) -> Arc<Adsr> {
+        let n = rack.num_modules();
+        controls[(n, 0)] = self.attack;
+        controls[(n, 1)] = self.decay;
+        controls[(n, 2)] = self.sustain;
+        controls[(n, 3)] = self.release;
+        controls[(n, 4)] = self.triggered;
+        let adsr = Arc::new(Adsr::new(n, self.ax, self.dx, self.rx));
+        rack.push(adsr.clone());
+        adsr
     }
 }
