+++ oscen-lib/src/lib.rs
@@ -25,32 +25,25 @@
 //! [`Signal`]: signal/trait.Signal.html
 //! [`Rack`]: signal/struct.Rack.html
 
-pub mod rack;
 /// A collection of some basic audio filters.
 pub mod filters;
-pub mod fil;
 /// An implementation of *freeverb*.
-pub mod reverb;
+// pub mod reverb;
 /// Envelope generators.
 pub mod envelopes;
-pub mod env;
 /// Core Oscen types and traits.
-pub mod signal;
+pub mod rack;
 /// Syth modules for combining other sytn modules.
 pub mod operators;
-pub mod ops;
 /// Some common (and some less common) oscillators.
 pub mod oscillators;
-pub mod osc;
 /// Wave shaping.
 pub mod shaping;
 /// Midi interface nodes.
 pub mod midi;
-pub mod mid;
 /// Utilites.
 pub mod utils;
-pub mod uti;
-/// Instruments.
-pub mod instruments;
-/// Sequencer
-pub mod sequencer;
+// Instruments.
+// pub mod instruments;
+// Sequencer
+// pub mod sequencer;
