+++ oscen-lib/src/operators.rs
@@ -1,477 +1,415 @@
-use super::oscillators::{ConstOsc, Oscillator, SignalFn};
-use super::signal::*;
-use super::utils::RingBuffer;
-use crate::{as_any_mut, std_signal};
-use std::any::Any;
-use std::ops::{Index, IndexMut};
-
-/// The `Union` module holds a vector of oscillators and plays one based on the
-/// active tag. The `level` field is used to set the volume of whichever signal
-/// is playing.
-#[derive(Clone)]
-pub struct Union {
-    tag: Tag,
-    waves: Vec<Tag>,
-    active: Tag,
-    level: In,
-    out: Real,
-}
-
-impl Union {
-    pub fn new(id_gen: &mut IdGen, waves: Vec<Tag>) -> Self {
-        let active = waves[0];
-        Union {
-            tag: id_gen.id(),
-            waves,
-            active,
-            level: 1.into(),
-            out: 0.0,
-        }
-    }
-
-    pub fn waves(&mut self, arg: Vec<Tag>) -> &mut Self {
-        self.waves = arg;
-        self
-    }
-
-    pub fn active(&mut self, arg: Tag) -> &mut Self {
-        self.active = arg;
-        self
-    }
-
-    pub fn level<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.level = arg.into();
-        self
-    }
-}
-
-impl Builder for Union {}
-
-impl Signal for Union {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, _sample_rate: Real) -> Real {
-        self.out = In::val(rack, self.level) * rack.output(self.active);
-        self.out
-    }
-}
-
-impl Index<usize> for Union {
-    type Output = Tag;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        &self.waves[index]
-    }
-}
-
-impl IndexMut<usize> for Union {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        &mut self.waves[index]
-    }
-}
-/// `Product` multiplies the signals of a vector of synth modules.
-#[derive(Clone)]
-pub struct Product {
-    tag: Tag,
-    waves: Vec<Tag>,
-    out: Real,
-}
-
-impl Product {
-    pub fn new(id_gen: &mut IdGen, waves: Vec<Tag>) -> Self {
-        Product {
-            tag: id_gen.id(),
-            waves,
-            out: 0.0,
-        }
-    }
-
-    pub fn waves(&mut self, arg: Vec<Tag>) -> &mut Self {
-        self.waves = arg;
-        self
-    }
-}
-
-impl Builder for Product {}
-
-impl Signal for Product {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, _sample_rate: Real) -> Real {
-        self.out = self.waves.iter().fold(1.0, |acc, n| acc * rack.output(*n));
-        self.out
-    }
-}
-
-impl Index<usize> for Product {
-    type Output = Tag;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        &self.waves[index]
-    }
-}
-
-impl IndexMut<usize> for Product {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        &mut self.waves[index]
-    }
-}
-
-/// "Voltage controlled amplifier" multiplies the volume of the `wave` by the
-/// value of `level`.
-#[derive(Copy, Clone)]
-pub struct Vca {
-    tag: Tag,
-    wave: Tag,
-    level: In,
-    out: Real,
-}
-
-impl Vca {
-    pub fn new(id_gen: &mut IdGen, wave: Tag) -> Self {
-        Self {
-            tag: id_gen.id(),
-            wave,
-            level: 1.into(),
-            out: 0.0,
-        }
-    }
-
-    pub fn wave(&mut self, arg: Tag) -> &mut Self {
-        self.wave = arg;
-        self
-    }
-
-    pub fn level<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.level = arg.into();
-        self
-    }
-}
-
-impl Builder for Vca {}
-
-impl Signal for Vca {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, _sample_rate: Real) -> Real {
-        self.out = rack.output(self.wave) * In::val(rack, self.level);
-        self.out
-    }
-}
-
-impl Index<&str> for Vca {
-    type Output = In;
-
-    fn index(&self, index: &str) -> &Self::Output {
-        match index {
-            "level" => &self.level,
-            _ => panic!("Vca does not have a field named: {}", index),
-        }
-    }
-}
-
-impl IndexMut<&str> for Vca {
-    fn index_mut(&mut self, index: &str) -> &mut Self::Output {
-        match index {
-            "level" => &mut self.level,
-            _ => panic!("Vca does not have a field named: {}", index),
-        }
-    }
-}
-
-/// Mixer with individual attenuverters for each wave plus an overall attenuverter.
-#[derive(Clone)]
+use crate::oscillators::{ConstBuilder, OscBuilder};
+use crate::rack::*;
+// use crate::utils::RingBuffer;
+use crate::{build, props, tag};
+use std::sync::Arc;
+#[derive(Debug, Clone)]
 pub struct Mixer {
     tag: Tag,
     waves: Vec<Tag>,
-    levels: Vec<In>,
-    level: In,
-    out: Real,
+}
+
+#[derive(Debug, Clone)]
+pub struct MixerBuilder {
+    waves: Vec<Tag>,
+}
+
+impl MixerBuilder {
+    pub fn new(waves: Vec<Tag>) -> Self {
+        Self { waves }
+    }
+    pub fn rack(&self, rack: &mut Rack) -> Arc<Mixer> {
+        let n = rack.num_modules();
+        let mix = Arc::new(Mixer::new(n.into(), self.waves.clone()));
+        rack.push(mix.clone());
+        mix
+    }
 }
 
 impl Mixer {
-    pub fn new(id_gen: &mut IdGen, waves: Vec<Tag>) -> Self {
-        let levels = waves.iter().map(|_| 1.into()).collect();
-        Mixer {
-            tag: id_gen.id(),
-            waves,
-            levels,
-            level: 1.into(),
-            out: 0.0,
-        }
-    }
-
-    pub fn waves(&mut self, arg: Vec<Tag>) -> &mut Self {
-        self.waves = arg;
-        self.levels.resize_with(self.waves.len(), || 0.5.into());
-        self
-    }
-
-    pub fn levels<T: Into<In>>(&mut self, arg: Vec<T>) -> &mut Self {
-        assert_eq!(
-            arg.len(),
-            self.waves.len(),
-            "Levels must have same length as waves"
-        );
-        let v = arg.into_iter().map(|x| x.into());
-        self.levels = v.collect();
-        self
-    }
-
-    pub fn level<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.level = arg.into();
-        self
-    }
-
-    pub fn level_nth<T: Into<In>>(&mut self, n: usize, arg: T) -> &mut Self {
-        self.levels[n] = arg.into();
-        self
+    fn new(tag: Tag, waves: Vec<Tag>) -> Self {
+        Self { tag, waves }
     }
 }
-
-impl Builder for Mixer {}
 
 impl Signal for Mixer {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, _sample_rate: Real) -> Real {
-        self.out = self.waves.iter().enumerate().fold(0.0, |acc, (i, n)| {
-            acc + rack.output(*n) * In::val(rack, self.levels[i])
-        }) * In::val(rack, self.level);
-        self.out
+    tag!();
+    fn signal(
+        &self,
+        _controls: &Controls,
+        _state: &mut State,
+        outputs: &mut Outputs,
+        _sample_rate: f32,
+    ) {
+        let out = self.waves.iter().fold(0.0, |acc, n| acc + outputs[(*n, 0)]);
+        outputs[(self.tag, 0)] = out;
     }
 }
 
-impl Index<usize> for Mixer {
-    type Output = Tag;
+#[derive(Debug, Clone)]
+pub struct Union {
+    tag: Tag,
+    waves: Vec<Tag>,
+}
 
-    fn index(&self, index: usize) -> &Self::Output {
-        &self.waves[index]
+#[derive(Clone)]
+pub struct UnionBuilder {
+    waves: Vec<Tag>,
+    active: Control,
+}
+
+impl UnionBuilder {
+    pub fn new(waves: Vec<Tag>) -> Self {
+        Self {
+            waves,
+            active: Control::I(0),
+        }
+    }
+    build!(active);
+    pub fn rack(&self, rack: &mut Rack, controls: &mut Controls) -> Arc<Union> {
+        let n = rack.num_modules();
+        controls[(n, 0)] = self.active;
+        let u = Arc::new(Union::new(n.into(), self.waves.clone()));
+        rack.push(u.clone());
+        u
     }
 }
 
-impl IndexMut<usize> for Mixer {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        &mut self.waves[index]
+impl Union {
+    pub fn new(tag: Tag, waves: Vec<Tag>) -> Self {
+        Self { tag, waves }
+    }
+    pub fn active(&self, controls: &Controls, outputs: &Outputs) -> usize {
+        let inp = controls[(self.tag, 0)];
+        outputs.integer(inp).expect("active must be Control::I")
+    }
+    pub fn set_active(&self, controls: &mut Controls, value: Control) {
+        controls[(self.tag, 0)] = value;
     }
 }
 
-/// A cross fade synth module, alpha = 0 means 100% wave 1.
+impl Signal for Union {
+    tag!();
+    fn signal(
+        &self,
+        controls: &Controls,
+        _state: &mut State,
+        outputs: &mut Outputs,
+        _sample_rate: f32,
+    ) {
+        let idx = self.active(controls, outputs);
+        let wave = self.waves[idx];
+        outputs[(self.tag, 0)] = outputs[(wave, 0)];
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct Product {
+    tag: Tag,
+    waves: Vec<Tag>,
+}
+
+#[derive(Debug, Clone)]
+pub struct ProductBuilder {
+    waves: Vec<Tag>,
+}
+
+impl ProductBuilder {
+    pub fn new(waves: Vec<Tag>) -> Self {
+        Self { waves }
+    }
+    pub fn rack(&self, rack: &mut Rack) -> Arc<Product> {
+        let n = rack.num_modules();
+        let p = Arc::new(Product::new(n.into(), self.waves.clone()));
+        rack.push(p.clone());
+        p
+    }
+}
+
+impl Product {
+    fn new(tag: Tag, waves: Vec<Tag>) -> Self {
+        Self { tag, waves }
+    }
+}
+
+impl Signal for Product {
+    tag!();
+    fn signal(
+        &self,
+        _controls: &Controls,
+        _state: &mut State,
+        outputs: &mut Outputs,
+        _sample_rate: f32,
+    ) {
+        let out = self.waves.iter().fold(1.0, |acc, n| acc * outputs[(*n, 0)]);
+        outputs[(self.tag, 0)] = out;
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
+pub struct Vca {
+    tag: Tag,
+    wave: Tag,
+}
+
+impl Vca {
+    pub fn new(tag: Tag, wave: Tag) -> Self {
+        Self { tag, wave }
+    }
+    props!(level, set_level, 0);
+}
+
+impl Signal for Vca {
+    tag!();
+    fn signal(
+        &self,
+        controls: &Controls,
+        _state: &mut State,
+        outputs: &mut Outputs,
+        _sample_rate: f32,
+    ) {
+        outputs[(self.tag, 0)] = self.level(controls, outputs) * outputs[(self.wave, 0)];
+    }
+}
+
 #[derive(Copy, Clone)]
+pub struct VcaBuilder {
+    wave: Tag,
+    level: Control,
+}
+
+impl VcaBuilder {
+    pub fn new(wave: Tag) -> Self {
+        Self {
+            wave,
+            level: 1.into(),
+        }
+    }
+    build!(level);
+    pub fn rack(&self, rack: &mut Rack, controls: &mut Controls) -> Arc<Vca> {
+        let n = rack.num_modules();
+        controls[(n, 0)] = self.level;
+        let vca = Arc::new(Vca::new(n.into(), self.wave));
+        rack.push(vca.clone());
+        vca
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
 pub struct CrossFade {
     tag: Tag,
-    wave1: In,
-    wave2: In,
-    alpha: In,
-    out: Real,
+    wave1: Tag,
+    wave2: Tag,
 }
 
 impl CrossFade {
-    pub fn new(id_gen: &mut IdGen, wave1: Tag, wave2: Tag) -> Self {
-        CrossFade {
-            tag: id_gen.id(),
-            wave1: wave1.into(),
-            wave2: wave2.into(),
-            alpha: (0.5).into(),
-            out: 0.0,
-        }
+    pub fn new(tag: Tag, wave1: Tag, wave2: Tag) -> Self {
+        Self { tag, wave1, wave2 }
     }
-
-    pub fn wave1<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.wave1 = arg.into();
-        self
-    }
-
-    pub fn wave2<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.wave2 = arg.into();
-        self
-    }
-
-    pub fn alpha<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.alpha = arg.into();
-        self
-    }
+    props!(alpha, set_alpha, 0);
 }
-
-impl Builder for CrossFade {}
 
 impl Signal for CrossFade {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, _sample_rate: Real) -> Real {
-        let alpha = In::val(rack, self.alpha);
-        self.out = alpha * In::val(rack, self.wave2) + (1.0 - alpha) * In::val(rack, self.wave1);
-        self.out
+    tag!();
+    fn signal(
+        &self,
+        controls: &Controls,
+        _state: &mut State,
+        outputs: &mut Outputs,
+        _sample_rate: f32,
+    ) {
+        let alpha = self.alpha(controls, outputs);
+        outputs[(self.tag, 0)] =
+            alpha * outputs[(self.wave2, 0)] + (1.0 - alpha) * outputs[(self.wave1, 0)];
     }
 }
 
-impl Index<&str> for CrossFade {
-    type Output = In;
+#[derive(Debug, Copy, Clone)]
+pub struct CrossFadeBuilder {
+    wave1: Tag,
+    wave2: Tag,
+    alpha: Control,
+}
 
-    fn index(&self, index: &str) -> &Self::Output {
-        match index {
-            "wave1" => &self.wave1,
-            "wave2" => &self.wave2,
-            "alpha" => &self.alpha,
-            _ => panic!("CrossFade does not have a field named: {}", index),
+impl CrossFadeBuilder {
+    pub fn new(wave1: Tag, wave2: Tag) -> Self {
+        Self {
+            wave1,
+            wave2,
+            alpha: 0.5.into(),
         }
     }
-}
-
-impl IndexMut<&str> for CrossFade {
-    fn index_mut(&mut self, index: &str) -> &mut Self::Output {
-        match index {
-            "wave1" => &mut self.wave1,
-            "wave2" => &mut self.wave2,
-            "alpha" => &mut self.alpha,
-            _ => panic!("CrossFade does not have a field named: {}", index),
-        }
+    build!(alpha);
+    pub fn rack(&self, rack: &mut Rack, controls: &mut Controls) -> Arc<CrossFade> {
+        let n = rack.num_modules();
+        controls[(n, 0)] = self.alpha;
+        let cf = Arc::new(CrossFade::new(n.into(), self.wave1, self.wave2));
+        rack.push(cf.clone());
+        cf
     }
 }
 
-/// A `Modulator` is designed to be the input to the `hz` field of a carrier
-/// wave. It takes control of the carriers frequency and modulates it's base
-/// hz by adding mod_idx * mod_hz * output of modulator wave.
 #[derive(Clone)]
 pub struct Modulator {
     tag: Tag,
-    wave: ArcMutex<Oscillator>,
-    hz: In,
-    ratio: In,
-    index: In,
-    hz_osc: ArcMutex<ConstOsc>,
-    ratio_osc: ArcMutex<ConstOsc>,
-    index_osc: ArcMutex<ConstOsc>,
-    mod_hz: ArcMutex<Product>,
-    amp_factor: ArcMutex<Product>,
-    mod_amp: ArcMutex<Mixer>,
-    carrier_hz: ArcMutex<Mixer>,
-    rack: Rack,
-    out: Real,
+    hz_tag: Tag,
+    ratio_tag: Tag,
+    index_tag: Tag,
 }
 
 impl Modulator {
-    pub fn new<H, R, I>(id_gen: &mut IdGen, signal_fn: SignalFn, hz: H, ratio: R, index: I) -> Self
-    where
-        H: Into<In> + Copy,
-        R: Into<In> + Copy,
-        I: Into<In> + Copy,
-    {
-        let mut rack = Rack::new();
-        let mut id = IdGen::new();
-        let hz_osc = ConstOsc::new(&mut id, hz.into()).rack(&mut rack);
-        let ratio_osc = ConstOsc::new(&mut id, ratio.into()).rack(&mut rack);
-        let index_osc = ConstOsc::new(&mut id, index.into()).rack(&mut rack);
-        let mod_hz = Product::new(&mut id, vec![ratio_osc.tag(), hz_osc.tag()]).rack(&mut rack);
-        let amp_factor = Product::new(
-            &mut id,
-            vec![index_osc.tag(), hz_osc.tag(), ratio_osc.tag()],
-        )
-        .rack(&mut rack);
-        let mod_amp = Mixer::new(&mut id, vec![hz_osc.tag(), amp_factor.tag()]).rack(&mut rack);
-        let wave = Oscillator::new(&mut id, signal_fn)
-            .hz(mod_hz.tag())
-            .amplitude(mod_amp.tag())
-            .rack(&mut rack);
-        let carrier_hz = Mixer::new(&mut id, vec![wave.tag(), hz_osc.tag()]).rack(&mut rack);
-        Modulator {
-            tag: id_gen.id(),
-            wave,
-            hz: hz.into(),
-            /// modulator frequency / carrier frequency
-            ratio: ratio.into(),
-            index: index.into(),
-            hz_osc,
-            ratio_osc,
-            index_osc,
-            mod_hz,
-            amp_factor,
-            mod_amp,
-            carrier_hz,
-            rack,
-            out: 0.0,
+    pub fn new(tag: Tag, hz_tag: Tag, ratio_tag: Tag, index_tag: Tag) -> Self {
+        Self {
+            tag,
+            hz_tag,
+            ratio_tag,
+            index_tag,
         }
     }
+    pub fn hz(&self, controls: &Controls, outputs: &Outputs) -> f32 {
+        let inp = controls[(self.hz_tag, 0)];
+        outputs.value(inp).unwrap()
+    }
+    pub fn set_hz(&self, controls: &mut Controls, value: Control) {
+        controls[(self.hz_tag, 0)] = value;
+    }
+    pub fn ratio(&self, controls: &Controls, outputs: &Outputs) -> f32 {
+        let inp = controls[(self.ratio_tag, 0)];
+        outputs.value(inp).unwrap()
+    }
+    pub fn set_ratio(&self, controls: &mut Controls, value: Control) {
+        controls[(self.ratio_tag, 0)] = value;
+    }
+    pub fn index(&self, controls: &Controls, outputs: &Outputs) -> f32 {
+        let inp = controls[(self.index_tag, 0)];
+        outputs.value(inp).unwrap()
+    }
+    pub fn set_index(&self, controls: &mut Controls, value: Control) {
+        controls[(self.index_tag, 0)] = value;
+    }
 }
-
-impl Builder for Modulator {}
 
 impl Signal for Modulator {
-    std_signal!();
-    fn signal(&mut self, _rack: &Rack, sample_rate: Real) -> Real {
-        self.out = self.rack.signal(sample_rate);
-        self.out
+    tag!();
+    fn signal(
+        &self,
+        _controls: &Controls,
+        _state: &mut State,
+        _outputs: &mut Outputs,
+        _sample_rate: f32,
+    ) {
     }
 }
 
-impl Index<&str> for Modulator {
-    type Output = In;
+#[derive(Debug, Clone)]
+pub struct ModulatorBuilder {
+    hz: Control,
+    ratio: Control,
+    index: Control,
+    signal_fn: SignalFn,
+}
 
-    fn index(&self, index: &str) -> &Self::Output {
-        match index {
-            "hz" => &self.hz,
-            "ratio" => &self.ratio,
-            "index" => &self.index,
-            _ => panic!("Modulator only does not have a field named:  {}", index),
+impl ModulatorBuilder {
+    pub fn new(signal_fn: SignalFn) -> Self {
+        Self {
+            hz: 0.into(),
+            ratio: 1.into(),
+            index: 0.into(),
+            signal_fn,
         }
     }
-}
-
-impl IndexMut<&str> for Modulator {
-    fn index_mut(&mut self, index: &str) -> &mut Self::Output {
-        match index {
-            "hz" => &mut self.hz,
-            "ratio" => &mut self.ratio,
-            "index" => &mut self.index,
-            _ => panic!("Modulator only does not have a field named:  {}", index),
-        }
+    build!(hz);
+    build!(ratio);
+    build!(index);
+    pub fn rack(
+        &self,
+        rack: &mut Rack,
+        controls: &mut Controls,
+        state: &mut State,
+    ) -> Arc<Modulator> {
+        let hz = ConstBuilder::new(self.hz).rack(rack, controls);
+        let ratio = ConstBuilder::new(self.ratio).rack(rack, controls);
+        let index = ConstBuilder::new(self.index).rack(rack, controls);
+        let mod_hz = ProductBuilder::new(vec![hz.tag(), ratio.tag()]).rack(rack);
+        let mod_amp = ProductBuilder::new(vec![hz.tag(), ratio.tag(), index.tag()]).rack(rack);
+        // let mod_amp = MixerBuilder::new(vec![hz.tag(), amp_factor.tag()]).rack(rack);
+        let modulator = OscBuilder::new(self.signal_fn)
+            .amplitude(mod_amp.tag())
+            .hz(mod_hz.tag())
+            .rack(rack, controls, state);
+        let carrier_hz = MixerBuilder::new(vec![modulator.tag(), hz.tag()]).rack(rack);
+        Arc::new(Modulator::new(
+            carrier_hz.tag(),
+            hz.tag(),
+            ratio.tag(),
+            index.tag(),
+        ))
     }
 }
 
-/// A variable length delay line.
-#[derive(Clone)]
+/*
 pub struct Delay {
     tag: Tag,
     wave: Tag,
-    delay_time: In,
-    ring_buffer: RingBuffer<Real>,
-    out: Real,
 }
 
 impl Delay {
-    pub fn new(id_gen: &mut IdGen, wave: Tag, delay_time: In) -> Self {
-        let ring = RingBuffer::<Real>::new(0.0, 0);
-        Self {
-            tag: id_gen.id(),
+    pub fn new<T: Into<Tag>>(tag: T, wave: Tag) -> Self {
+        Delay {
+            tag: tag.into(),
             wave,
-            delay_time,
-            ring_buffer: ring,
-            out: 0.0,
         }
     }
-
-    pub fn wave(&mut self, arg: Tag) -> &mut Self {
-        self.wave = arg;
-        self
-    }
-
-    pub fn delay_time<T: Into<In>>(&mut self, arg: T) -> &mut Self {
-        self.delay_time = arg.into();
-        self
-    }
+    props!(delay_time, set_delay_time, 0);
 }
 
-impl Builder for Delay {}
-
 impl Signal for Delay {
-    std_signal!();
-    fn signal(&mut self, rack: &Rack, sample_rate: Real) -> Real {
-        let delay = In::val(rack, self.delay_time) * sample_rate;
+    tag!();
+    fn signal(
+        &self,
+        controls: &Controls,
+        _state: &mut State,
+        outputs: &mut Outputs,
+        sample_rate: f32,
+    ) {
+        let delay = self.delay_time(controls, outputs) * sample_rate;
         let rp = self.ring_buffer.read_pos;
         let wp = (delay + rp).ceil();
         self.ring_buffer.set_write_pos(wp as usize);
-        self.ring_buffer.set_read_pos(wp - delay);
-        if delay > self.ring_buffer.len() as Real - 3.0 {
-            self.ring_buffer.resize(delay as usize + 3);
+        self.ring_buffer.set_read_pos(rp - delay);
+        if delay > self.ring_buffer.len() as f32 - 3.0 {
+            panic!("Ring buffer too small for dalay {}", delay);
         }
-        let val = rack.output(self.wave);
+        let val = outputs[(self.wave, 0)];
         self.ring_buffer.push(val);
-        self.out = self.ring_buffer.get_cubic();
-        self.out
+        outputs[(self.tag, 0)] = self.ring_buffer.get_cubic();
     }
 }
+
+pub struct DelayBuilder<'a> {
+    wave: Tag,
+    ring_buffer: &'a mut RingBuffer<'a, f32>,
+    delay_time: Control,
+}
+
+impl<'a> DelayBuilder<'a> {
+    pub fn new(wave: Tag, ring_buffer: &'a mut RingBuffer<'a, f32>, delay_time: Control) -> Self {
+        Self {
+            wave,
+            ring_buffer,
+            delay_time,
+        }
+    }
+    build!(delay_time);
+    pub fn rack(&'static mut self, rack: &mut Rack, controls: &mut Controls) -> Arc<Delay> {
+        let n = rack.num_modules();
+        controls[(n, 0)] = self.delay_time;
+        let wave = self.wave;
+        // let mut ring_buffer = &*self.ring_buffer;
+        let ring_buffer = &mut *self.ring_buffer;
+        let delay = Arc::new(Delay::new(n, wave));
+        rack.push(delay);
+        delay
+    }
+}
+*/
